feat: reicepts component created
